chore: rename testing from .ts to .mts
