+++ src/__tests__/axios.spec.mts
@@ -1,10 +1,10 @@
-import Wreck from '@hapi/wreck';
 import test from 'ava';
+import axios from 'axios';
 import { CookieJar } from 'tough-cookie';
 
-import { HttpCookieAgent } from '../';
+import { HttpCookieAgent } from '../index.js';
 
-import { createTestServer, readStream } from './helpers';
+import { createTestServer, readStream } from './helpers.mjs';
 
 test('should set cookies to CookieJar from Set-Cookie header', async (t) => {
   const jar = new CookieJar();
@@ -17,8 +17,8 @@ test('should set cookies to CookieJar from Set-Cookie header', async (t) => {
     },
   ]);
 
-  await Wreck.get(`http://localhost:${port}`, {
-    agent,
+  await axios.get(`http://localhost:${port}`, {
+    httpAgent: agent,
   });
 
   const cookies = await jar.getCookies(`http://localhost:${port}`);
@@ -39,8 +39,8 @@ test('should set cookies to CookieJar from multiple Set-Cookie headers', async (
     },
   ]);
 
-  await Wreck.get(`http://localhost:${port}`, {
-    agent,
+  await axios.get(`http://localhost:${port}`, {
+    httpAgent: agent,
   });
 
   const cookies = await jar.getCookies(`http://localhost:${port}`);
@@ -64,8 +64,8 @@ test('should send cookies from CookieJar', async (t) => {
 
   await jar.setCookie('key=value', `http://localhost:${port}`);
 
-  await Wreck.get(`http://localhost:${port}`, {
-    agent,
+  await axios.get(`http://localhost:${port}`, {
+    httpAgent: agent,
   });
 
   t.plan(1);
@@ -84,9 +84,9 @@ test('should send cookies from both a request options and CookieJar', async (t) 
 
   await jar.setCookie('key1=value1', `http://localhost:${port}`);
 
-  await Wreck.get(`http://localhost:${port}`, {
-    agent,
+  await axios.get(`http://localhost:${port}`, {
     headers: { Cookie: 'key2=value2' },
+    httpAgent: agent,
   });
 
   t.plan(1);
@@ -105,9 +105,9 @@ test('should send cookies from a request options when the key is duplicated in b
 
   await jar.setCookie('key=notexpected', `http://localhost:${port}`);
 
-  await Wreck.get(`http://localhost:${port}`, {
-    agent,
+  await axios.get(`http://localhost:${port}`, {
     headers: { Cookie: 'key=expected' },
+    httpAgent: agent,
   });
 
   t.plan(1);
@@ -130,9 +130,8 @@ test('should send cookies from the first response when redirecting', async (t) =
     },
   ]);
 
-  await Wreck.get(`http://localhost:${port}`, {
-    agent,
-    redirects: 1,
+  await axios.get(`http://localhost:${port}`, {
+    httpAgent: agent,
   });
 
   t.plan(1);
@@ -153,8 +152,8 @@ test('should emit error when CookieJar#getCookies throws error.', async (t) => {
   ]);
 
   await t.throwsAsync(() => {
-    return Wreck.get(`http://localhost:${port}`, {
-      agent,
+    return axios.get(`http://localhost:${port}`, {
+      httpAgent: agent,
     });
   });
 
@@ -175,9 +174,9 @@ test('should emit error when CookieJar#setCookie throws error.', async (t) => {
     },
   ]);
 
-  await t.throwsAsync(async () => {
-    return Wreck.get(`http://localhost:${port}`, {
-      agent,
+  await t.throwsAsync(() => {
+    return axios.get(`http://localhost:${port}`, {
+      httpAgent: agent,
     });
   });
 
@@ -203,9 +202,8 @@ test('should send post data when keepalive is enabled', async (t) => {
   await jar.setCookie('key=expected', `http://localhost:${port}`);
 
   for (let idx = 0; idx < times; idx++) {
-    await Wreck.post(`http://localhost:${port}`, {
-      agent,
-      payload: `{ "index": "${idx}" }`,
+    await axios.post(`http://localhost:${port}`, `{ "index": "${idx}" }`, {
+      httpAgent: agent,
     });
   }
