+++ src/__tests__/got.spec.mts
@@ -1,10 +1,10 @@
 import test from 'ava';
-import axios from 'axios';
+import got from 'got';
 import { CookieJar } from 'tough-cookie';
 
-import { HttpCookieAgent } from '../';
+import { HttpCookieAgent } from '../index.js';
 
-import { createTestServer, readStream } from './helpers';
+import { createTestServer, readStream } from './helpers.mjs';
 
 test('should set cookies to CookieJar from Set-Cookie header', async (t) => {
   const jar = new CookieJar();
@@ -17,8 +17,8 @@ test('should set cookies to CookieJar from Set-Cookie header', async (t) => {
     },
   ]);
 
-  await axios.get(`http://localhost:${port}`, {
-    httpAgent: agent,
+  await got(`http://localhost:${port}`, {
+    agent: { http: agent },
   });
 
   const cookies = await jar.getCookies(`http://localhost:${port}`);
@@ -39,8 +39,8 @@ test('should set cookies to CookieJar from multiple Set-Cookie headers', async (
     },
   ]);
 
-  await axios.get(`http://localhost:${port}`, {
-    httpAgent: agent,
+  await got(`http://localhost:${port}`, {
+    agent: { http: agent },
   });
 
   const cookies = await jar.getCookies(`http://localhost:${port}`);
@@ -64,8 +64,8 @@ test('should send cookies from CookieJar', async (t) => {
 
   await jar.setCookie('key=value', `http://localhost:${port}`);
 
-  await axios.get(`http://localhost:${port}`, {
-    httpAgent: agent,
+  await got(`http://localhost:${port}`, {
+    agent: { http: agent },
   });
 
   t.plan(1);
@@ -84,9 +84,9 @@ test('should send cookies from both a request options and CookieJar', async (t) 
 
   await jar.setCookie('key1=value1', `http://localhost:${port}`);
 
-  await axios.get(`http://localhost:${port}`, {
+  await got(`http://localhost:${port}`, {
+    agent: { http: agent },
     headers: { Cookie: 'key2=value2' },
-    httpAgent: agent,
   });
 
   t.plan(1);
@@ -105,9 +105,9 @@ test('should send cookies from a request options when the key is duplicated in b
 
   await jar.setCookie('key=notexpected', `http://localhost:${port}`);
 
-  await axios.get(`http://localhost:${port}`, {
+  await got(`http://localhost:${port}`, {
+    agent: { http: agent },
     headers: { Cookie: 'key=expected' },
-    httpAgent: agent,
   });
 
   t.plan(1);
@@ -130,8 +130,8 @@ test('should send cookies from the first response when redirecting', async (t) =
     },
   ]);
 
-  await axios.get(`http://localhost:${port}`, {
-    httpAgent: agent,
+  await got(`http://localhost:${port}`, {
+    agent: { http: agent },
   });
 
   t.plan(1);
@@ -152,8 +152,9 @@ test('should emit error when CookieJar#getCookies throws error.', async (t) => {
   ]);
 
   await t.throwsAsync(() => {
-    return axios.get(`http://localhost:${port}`, {
-      httpAgent: agent,
+    return got(`http://localhost:${port}`, {
+      agent: { http: agent },
+      retry: 0,
     });
   });
 
@@ -174,9 +175,10 @@ test('should emit error when CookieJar#setCookie throws error.', async (t) => {
     },
   ]);
 
-  await t.throwsAsync(() => {
-    return axios.get(`http://localhost:${port}`, {
-      httpAgent: agent,
+  await t.throwsAsync(async () => {
+    return got(`http://localhost:${port}`, {
+      agent: { http: agent },
+      retry: 0,
     });
   });
 
@@ -202,8 +204,11 @@ test('should send post data when keepalive is enabled', async (t) => {
   await jar.setCookie('key=expected', `http://localhost:${port}`);
 
   for (let idx = 0; idx < times; idx++) {
-    await axios.post(`http://localhost:${port}`, `{ "index": "${idx}" }`, {
-      httpAgent: agent,
+    await got(`http://localhost:${port}`, {
+      agent: { http: agent },
+      body: `{ "index": "${idx}" }`,
+      method: 'POST',
+      retry: 0,
     });
   }
