+++ src/__tests__/mixed_cookie_agent.spec.mts
@@ -1,14 +1,32 @@
+import http from 'node:http';
+
 import test from 'ava';
-import phin from 'phin';
 import { CookieJar } from 'tough-cookie';
 
-import { HttpCookieAgent } from '../';
+import { MixedCookieAgent } from '../index.js';
 
-import { createTestServer, readStream } from './helpers';
+import { createTestServer } from './helpers.mjs';
+
+export function request(url: string, options: http.RequestOptions) {
+  const req = http.request(url, options);
+
+  const promise = new Promise<http.IncomingMessage>((resolve, reject) => {
+    req.on('response', (res) => {
+      res.on('error', (err) => reject(err));
+      // eslint-disable-next-line @typescript-eslint/no-empty-function
+      res.on('data', () => {});
+      res.on('end', () => resolve(res));
+    });
+    req.on('error', (err) => reject(err));
+  });
+  req.end();
+
+  return { promise, req };
+}
 
 test('should set cookies to CookieJar from Set-Cookie header', async (t) => {
   const jar = new CookieJar();
-  const agent = new HttpCookieAgent({ jar });
+  const agent = new MixedCookieAgent({ jar });
 
   const { port } = await createTestServer([
     (_req, res) => {
@@ -17,10 +35,11 @@ test('should set cookies to CookieJar from Set-Cookie header', async (t) => {
     },
   ]);
 
-  await phin({
-    core: { agent },
-    url: `http://localhost:${port}`,
+  const { promise } = request(`http://localhost:${port}`, {
+    agent,
+    method: 'GET',
   });
+  await promise;
 
   const cookies = await jar.getCookies(`http://localhost:${port}`);
   t.is(cookies.length, 1);
@@ -31,7 +50,7 @@ test('should set cookies to CookieJar from Set-Cookie header', async (t) => {
 
 test('should set cookies to CookieJar from multiple Set-Cookie headers', async (t) => {
   const jar = new CookieJar();
-  const agent = new HttpCookieAgent({ jar });
+  const agent = new MixedCookieAgent({ jar });
 
   const { port } = await createTestServer([
     (_req, res) => {
@@ -40,10 +59,11 @@ test('should set cookies to CookieJar from multiple Set-Cookie headers', async (
     },
   ]);
 
-  await phin({
-    core: { agent },
-    url: `http://localhost:${port}`,
+  const { promise } = request(`http://localhost:${port}`, {
+    agent,
+    method: 'GET',
   });
+  await promise;
 
   const cookies = await jar.getCookies(`http://localhost:${port}`);
   t.is(cookies.length, 2);
@@ -55,7 +75,7 @@ test('should set cookies to CookieJar from multiple Set-Cookie headers', async (
 
 test('should send cookies from CookieJar', async (t) => {
   const jar = new CookieJar();
-  const agent = new HttpCookieAgent({ jar });
+  const agent = new MixedCookieAgent({ jar });
 
   const { port } = await createTestServer([
     (req, res) => {
@@ -66,17 +86,18 @@ test('should send cookies from CookieJar', async (t) => {
 
   await jar.setCookie('key=value', `http://localhost:${port}`);
 
-  await phin({
-    core: { agent },
-    url: `http://localhost:${port}`,
+  const { promise } = request(`http://localhost:${port}`, {
+    agent,
+    method: 'GET',
   });
+  await promise;
 
   t.plan(1);
 });
 
 test('should send cookies from both a request options and CookieJar', async (t) => {
   const jar = new CookieJar();
-  const agent = new HttpCookieAgent({ jar });
+  const agent = new MixedCookieAgent({ jar });
 
   const { port } = await createTestServer([
     (req, res) => {
@@ -87,18 +108,19 @@ test('should send cookies from both a request options and CookieJar', async (t) 
 
   await jar.setCookie('key1=value1', `http://localhost:${port}`);
 
-  await phin({
-    core: { agent },
+  const { promise } = request(`http://localhost:${port}`, {
+    agent,
     headers: { Cookie: 'key2=value2' },
-    url: `http://localhost:${port}`,
+    method: 'GET',
   });
+  await promise;
 
   t.plan(1);
 });
 
 test('should send cookies from a request options when the key is duplicated in both a request options and CookieJar', async (t) => {
   const jar = new CookieJar();
-  const agent = new HttpCookieAgent({ jar });
+  const agent = new MixedCookieAgent({ jar });
 
   const { port } = await createTestServer([
     (req, res) => {
@@ -109,37 +131,12 @@ test('should send cookies from a request options when the key is duplicated in b
 
   await jar.setCookie('key=notexpected', `http://localhost:${port}`);
 
-  await phin({
-    core: { agent },
+  const { promise } = request(`http://localhost:${port}`, {
+    agent,
     headers: { Cookie: 'key=expected' },
-    url: `http://localhost:${port}`,
+    method: 'GET',
   });
-
-  t.plan(1);
-});
-
-test('should send cookies from the first response when redirecting', async (t) => {
-  const jar = new CookieJar();
-  const agent = new HttpCookieAgent({ jar });
-
-  const { port } = await createTestServer([
-    (_req, res) => {
-      res.statusCode = 301;
-      res.setHeader('Location', '/redirect');
-      res.setHeader('Set-Cookie', 'key=value');
-      res.end();
-    },
-    (req, res) => {
-      t.is(req.headers['cookie'], 'key=value');
-      res.end();
-    },
-  ]);
-
-  await phin({
-    core: { agent },
-    followRedirects: true,
-    url: `http://localhost:${port}`,
-  });
+  await promise;
 
   t.plan(1);
 });
@@ -149,7 +146,7 @@ test('should emit error when CookieJar#getCookies throws error.', async (t) => {
   jar.getCookies = async () => {
     throw new Error('Error');
   };
-  const agent = new HttpCookieAgent({ jar });
+  const agent = new MixedCookieAgent({ jar });
 
   const { port } = await createTestServer([
     (_req, res) => {
@@ -158,12 +155,11 @@ test('should emit error when CookieJar#getCookies throws error.', async (t) => {
     },
   ]);
 
-  await t.throwsAsync(() => {
-    return phin({
-      core: { agent },
-      url: `http://localhost:${port}`,
-    });
+  const { promise } = request(`http://localhost:${port}`, {
+    agent,
+    method: 'GET',
   });
+  await t.throwsAsync(() => promise);
 
   t.plan(1);
 });
@@ -173,7 +169,7 @@ test('should emit error when CookieJar#setCookie throws error.', async (t) => {
   jar.setCookie = async () => {
     throw new Error('Error');
   };
-  const agent = new HttpCookieAgent({ jar });
+  const agent = new MixedCookieAgent({ jar });
 
   const { port } = await createTestServer([
     (_req, res) => {
@@ -182,42 +178,11 @@ test('should emit error when CookieJar#setCookie throws error.', async (t) => {
     },
   ]);
 
-  await t.throwsAsync(async () => {
-    return phin({
-      core: { agent },
-      url: `http://localhost:${port}`,
-    });
+  const { promise } = request(`http://localhost:${port}`, {
+    agent,
+    method: 'GET',
   });
+  await t.throwsAsync(() => promise);
 
   t.plan(1);
-});
-
-test('should send post data when keepalive is enabled', async (t) => {
-  const times = 2;
-
-  const jar = new CookieJar();
-  const agent = new HttpCookieAgent({ jar, keepAlive: true });
-
-  const { port } = await createTestServer(
-    Array.from({ length: times }, (_, idx) => {
-      return async (req, res) => {
-        t.is(await readStream(req), `{ "index": "${idx}" }`);
-        t.is(req.headers['cookie'], 'key=expected');
-        res.end();
-      };
-    }),
-  );
-
-  await jar.setCookie('key=expected', `http://localhost:${port}`);
-
-  for (let idx = 0; idx < times; idx++) {
-    await phin({
-      core: { agent },
-      data: `{ "index": "${idx}" }`,
-      method: 'POST',
-      url: `http://localhost:${port}`,
-    });
-  }
-
-  t.plan(times * 2);
 });
