+++ src/__tests__/phin.spec.mts
@@ -1,27 +1,10 @@
-import http from 'node:http';
-
 import test from 'ava';
+import phin from 'phin';
 import { CookieJar } from 'tough-cookie';
 
-import { HttpCookieAgent } from '../';
+import { HttpCookieAgent } from '../index.js';
 
-import { createTestServer, readStream } from './helpers';
-
-export function request(url: string, options: http.RequestOptions) {
-  const req = http.request(url, options);
-
-  const promise = new Promise<http.IncomingMessage>((resolve, reject) => {
-    req.on('response', (res) => {
-      res.on('error', (err) => reject(err));
-      // eslint-disable-next-line @typescript-eslint/no-empty-function
-      res.on('data', () => {});
-      res.on('end', () => resolve(res));
-    });
-    req.on('error', (err) => reject(err));
-  });
-
-  return { promise, req };
-}
+import { createTestServer, readStream } from './helpers.mjs';
 
 test('should set cookies to CookieJar from Set-Cookie header', async (t) => {
   const jar = new CookieJar();
@@ -34,12 +17,10 @@ test('should set cookies to CookieJar from Set-Cookie header', async (t) => {
     },
   ]);
 
-  const { promise, req } = request(`http://localhost:${port}`, {
-    agent,
-    method: 'GET',
+  await phin({
+    core: { agent },
+    url: `http://localhost:${port}`,
   });
-  req.end();
-  await promise;
 
   const cookies = await jar.getCookies(`http://localhost:${port}`);
   t.is(cookies.length, 1);
@@ -59,12 +40,10 @@ test('should set cookies to CookieJar from multiple Set-Cookie headers', async (
     },
   ]);
 
-  const { promise, req } = request(`http://localhost:${port}`, {
-    agent,
-    method: 'GET',
+  await phin({
+    core: { agent },
+    url: `http://localhost:${port}`,
   });
-  req.end();
-  await promise;
 
   const cookies = await jar.getCookies(`http://localhost:${port}`);
   t.is(cookies.length, 2);
@@ -87,35 +66,10 @@ test('should send cookies from CookieJar', async (t) => {
 
   await jar.setCookie('key=value', `http://localhost:${port}`);
 
-  const { promise, req } = request(`http://localhost:${port}`, {
-    agent,
-    method: 'GET',
+  await phin({
+    core: { agent },
+    url: `http://localhost:${port}`,
   });
-  req.end();
-  await promise;
-
-  t.plan(1);
-});
-
-test('should send cookies from CookieJar when value is url-encoded', async (t) => {
-  const jar = new CookieJar();
-  const agent = new HttpCookieAgent({ jar });
-
-  const { port } = await createTestServer([
-    (req, res) => {
-      t.is(req.headers['cookie'], 'key=hello%20world');
-      res.end();
-    },
-  ]);
-
-  await jar.setCookie('key=hello%20world', `http://localhost:${port}`);
-
-  const { promise, req } = request(`http://localhost:${port}`, {
-    agent,
-    method: 'GET',
-  });
-  req.end();
-  await promise;
 
   t.plan(1);
 });
@@ -133,13 +87,11 @@ test('should send cookies from both a request options and CookieJar', async (t) 
 
   await jar.setCookie('key1=value1', `http://localhost:${port}`);
 
-  const { promise, req } = request(`http://localhost:${port}`, {
-    agent,
+  await phin({
+    core: { agent },
     headers: { Cookie: 'key2=value2' },
-    method: 'GET',
+    url: `http://localhost:${port}`,
   });
-  req.end();
-  await promise;
 
   t.plan(1);
 });
@@ -157,13 +109,37 @@ test('should send cookies from a request options when the key is duplicated in b
 
   await jar.setCookie('key=notexpected', `http://localhost:${port}`);
 
-  const { promise, req } = request(`http://localhost:${port}`, {
-    agent,
+  await phin({
+    core: { agent },
     headers: { Cookie: 'key=expected' },
-    method: 'GET',
+    url: `http://localhost:${port}`,
   });
-  req.end();
-  await promise;
+
+  t.plan(1);
+});
+
+test('should send cookies from the first response when redirecting', async (t) => {
+  const jar = new CookieJar();
+  const agent = new HttpCookieAgent({ jar });
+
+  const { port } = await createTestServer([
+    (_req, res) => {
+      res.statusCode = 301;
+      res.setHeader('Location', '/redirect');
+      res.setHeader('Set-Cookie', 'key=value');
+      res.end();
+    },
+    (req, res) => {
+      t.is(req.headers['cookie'], 'key=value');
+      res.end();
+    },
+  ]);
+
+  await phin({
+    core: { agent },
+    followRedirects: true,
+    url: `http://localhost:${port}`,
+  });
 
   t.plan(1);
 });
@@ -182,12 +158,12 @@ test('should emit error when CookieJar#getCookies throws error.', async (t) => {
     },
   ]);
 
-  const { promise, req } = request(`http://localhost:${port}`, {
-    agent,
-    method: 'GET',
+  await t.throwsAsync(() => {
+    return phin({
+      core: { agent },
+      url: `http://localhost:${port}`,
+    });
   });
-  req.end();
-  await t.throwsAsync(() => promise);
 
   t.plan(1);
 });
@@ -206,50 +182,12 @@ test('should emit error when CookieJar#setCookie throws error.', async (t) => {
     },
   ]);
 
-  const { promise, req } = request(`http://localhost:${port}`, {
-    agent,
-    method: 'GET',
+  await t.throwsAsync(async () => {
+    return phin({
+      core: { agent },
+      url: `http://localhost:${port}`,
+    });
   });
-  req.end();
-  await t.throwsAsync(() => promise);
-
-  t.plan(1);
-});
-
-test('should send cookies even when target is same host but different port', async (t) => {
-  const jar = new CookieJar();
-  const agent = new HttpCookieAgent({ jar });
-
-  const { port: firstServerPort } = await createTestServer([
-    (_req, res) => {
-      res.setHeader('Set-Cookie', 'key=expected');
-      res.end();
-    },
-  ]);
-
-  const { port: secondServerPort } = await createTestServer([
-    (req, res) => {
-      t.is(req.headers['cookie'], 'key=expected');
-      res.end();
-    },
-  ]);
-
-  {
-    const { promise, req } = request(`http://localhost:${firstServerPort}`, {
-      agent,
-      method: 'GET',
-    });
-    req.end();
-    await promise;
-  }
-  {
-    const { promise, req } = request(`http://localhost:${secondServerPort}`, {
-      agent,
-      method: 'GET',
-    });
-    req.end();
-    await promise;
-  }
 
   t.plan(1);
 });
@@ -273,12 +211,12 @@ test('should send post data when keepalive is enabled', async (t) => {
   await jar.setCookie('key=expected', `http://localhost:${port}`);
 
   for (let idx = 0; idx < times; idx++) {
-    const { promise, req } = request(`http://localhost:${port}`, {
-      agent,
+    await phin({
+      core: { agent },
+      data: `{ "index": "${idx}" }`,
       method: 'POST',
+      url: `http://localhost:${port}`,
     });
-    req.end(`{ "index": "${idx}" }`);
-    await promise;
   }
 
   t.plan(times * 2);
