+++ src/__tests__/superagent.spec.mts
@@ -1,10 +1,10 @@
 import test from 'ava';
-import got from 'got';
+import superagent from 'superagent';
 import { CookieJar } from 'tough-cookie';
 
-import { HttpCookieAgent } from '../';
+import { HttpCookieAgent } from '../index.js';
 
-import { createTestServer, readStream } from './helpers';
+import { createTestServer, readStream } from './helpers.mjs';
 
 test('should set cookies to CookieJar from Set-Cookie header', async (t) => {
   const jar = new CookieJar();
@@ -17,9 +17,7 @@ test('should set cookies to CookieJar from Set-Cookie header', async (t) => {
     },
   ]);
 
-  await got(`http://localhost:${port}`, {
-    agent: { http: agent },
-  });
+  await superagent.get(`http://localhost:${port}`).agent(agent);
 
   const cookies = await jar.getCookies(`http://localhost:${port}`);
   t.is(cookies.length, 1);
@@ -39,9 +37,7 @@ test('should set cookies to CookieJar from multiple Set-Cookie headers', async (
     },
   ]);
 
-  await got(`http://localhost:${port}`, {
-    agent: { http: agent },
-  });
+  await superagent.get(`http://localhost:${port}`).agent(agent);
 
   const cookies = await jar.getCookies(`http://localhost:${port}`);
   t.is(cookies.length, 2);
@@ -64,9 +60,7 @@ test('should send cookies from CookieJar', async (t) => {
 
   await jar.setCookie('key=value', `http://localhost:${port}`);
 
-  await got(`http://localhost:${port}`, {
-    agent: { http: agent },
-  });
+  await superagent.get(`http://localhost:${port}`).agent(agent);
 
   t.plan(1);
 });
@@ -84,10 +78,7 @@ test('should send cookies from both a request options and CookieJar', async (t) 
 
   await jar.setCookie('key1=value1', `http://localhost:${port}`);
 
-  await got(`http://localhost:${port}`, {
-    agent: { http: agent },
-    headers: { Cookie: 'key2=value2' },
-  });
+  await superagent.get(`http://localhost:${port}`).agent(agent).set('Cookie', 'key2=value2');
 
   t.plan(1);
 });
@@ -105,10 +96,7 @@ test('should send cookies from a request options when the key is duplicated in b
 
   await jar.setCookie('key=notexpected', `http://localhost:${port}`);
 
-  await got(`http://localhost:${port}`, {
-    agent: { http: agent },
-    headers: { Cookie: 'key=expected' },
-  });
+  await superagent.get(`http://localhost:${port}`).agent(agent).set('Cookie', 'key=expected');
 
   t.plan(1);
 });
@@ -130,9 +118,7 @@ test('should send cookies from the first response when redirecting', async (t) =
     },
   ]);
 
-  await got(`http://localhost:${port}`, {
-    agent: { http: agent },
-  });
+  await superagent.get(`http://localhost:${port}`).agent(agent);
 
   t.plan(1);
 });
@@ -152,10 +138,7 @@ test('should emit error when CookieJar#getCookies throws error.', async (t) => {
   ]);
 
   await t.throwsAsync(() => {
-    return got(`http://localhost:${port}`, {
-      agent: { http: agent },
-      retry: 0,
-    });
+    return superagent.get(`http://localhost:${port}`).agent(agent);
   });
 
   t.plan(1);
@@ -176,10 +159,7 @@ test('should emit error when CookieJar#setCookie throws error.', async (t) => {
   ]);
 
   await t.throwsAsync(async () => {
-    return got(`http://localhost:${port}`, {
-      agent: { http: agent },
-      retry: 0,
-    });
+    return superagent.get(`http://localhost:${port}`).agent(agent);
   });
 
   t.plan(1);
@@ -204,12 +184,7 @@ test('should send post data when keepalive is enabled', async (t) => {
   await jar.setCookie('key=expected', `http://localhost:${port}`);
 
   for (let idx = 0; idx < times; idx++) {
-    await got(`http://localhost:${port}`, {
-      agent: { http: agent },
-      body: `{ "index": "${idx}" }`,
-      method: 'POST',
-      retry: 0,
-    });
+    await superagent.post(`http://localhost:${port}`).send(`{ "index": "${idx}" }`).agent(agent);
   }
 
   t.plan(times * 2);
