+++ src/__tests__/urllib.spec.mts
@@ -1,32 +1,14 @@
-import http from 'node:http';
-
 import test from 'ava';
 import { CookieJar } from 'tough-cookie';
+import urllib from 'urllib';
 
-import { MixedCookieAgent } from '../';
+import { HttpCookieAgent } from '../index.js';
 
-import { createTestServer } from './helpers';
-
-export function request(url: string, options: http.RequestOptions) {
-  const req = http.request(url, options);
-
-  const promise = new Promise<http.IncomingMessage>((resolve, reject) => {
-    req.on('response', (res) => {
-      res.on('error', (err) => reject(err));
-      // eslint-disable-next-line @typescript-eslint/no-empty-function
-      res.on('data', () => {});
-      res.on('end', () => resolve(res));
-    });
-    req.on('error', (err) => reject(err));
-  });
-  req.end();
-
-  return { promise, req };
-}
+import { createTestServer, readStream } from './helpers.mjs';
 
 test('should set cookies to CookieJar from Set-Cookie header', async (t) => {
   const jar = new CookieJar();
-  const agent = new MixedCookieAgent({ jar });
+  const agent = new HttpCookieAgent({ jar });
 
   const { port } = await createTestServer([
     (_req, res) => {
@@ -35,11 +17,9 @@ test('should set cookies to CookieJar from Set-Cookie header', async (t) => {
     },
   ]);
 
-  const { promise } = request(`http://localhost:${port}`, {
+  await urllib.request(`http://localhost:${port}`, {
     agent,
-    method: 'GET',
   });
-  await promise;
 
   const cookies = await jar.getCookies(`http://localhost:${port}`);
   t.is(cookies.length, 1);
@@ -50,7 +30,7 @@ test('should set cookies to CookieJar from Set-Cookie header', async (t) => {
 
 test('should set cookies to CookieJar from multiple Set-Cookie headers', async (t) => {
   const jar = new CookieJar();
-  const agent = new MixedCookieAgent({ jar });
+  const agent = new HttpCookieAgent({ jar });
 
   const { port } = await createTestServer([
     (_req, res) => {
@@ -59,11 +39,9 @@ test('should set cookies to CookieJar from multiple Set-Cookie headers', async (
     },
   ]);
 
-  const { promise } = request(`http://localhost:${port}`, {
+  await urllib.request(`http://localhost:${port}`, {
     agent,
-    method: 'GET',
   });
-  await promise;
 
   const cookies = await jar.getCookies(`http://localhost:${port}`);
   t.is(cookies.length, 2);
@@ -75,7 +53,7 @@ test('should set cookies to CookieJar from multiple Set-Cookie headers', async (
 
 test('should send cookies from CookieJar', async (t) => {
   const jar = new CookieJar();
-  const agent = new MixedCookieAgent({ jar });
+  const agent = new HttpCookieAgent({ jar });
 
   const { port } = await createTestServer([
     (req, res) => {
@@ -86,18 +64,16 @@ test('should send cookies from CookieJar', async (t) => {
 
   await jar.setCookie('key=value', `http://localhost:${port}`);
 
-  const { promise } = request(`http://localhost:${port}`, {
+  await urllib.request(`http://localhost:${port}`, {
     agent,
-    method: 'GET',
   });
-  await promise;
 
   t.plan(1);
 });
 
 test('should send cookies from both a request options and CookieJar', async (t) => {
   const jar = new CookieJar();
-  const agent = new MixedCookieAgent({ jar });
+  const agent = new HttpCookieAgent({ jar });
 
   const { port } = await createTestServer([
     (req, res) => {
@@ -108,19 +84,17 @@ test('should send cookies from both a request options and CookieJar', async (t) 
 
   await jar.setCookie('key1=value1', `http://localhost:${port}`);
 
-  const { promise } = request(`http://localhost:${port}`, {
+  await urllib.request(`http://localhost:${port}`, {
     agent,
     headers: { Cookie: 'key2=value2' },
-    method: 'GET',
   });
-  await promise;
 
   t.plan(1);
 });
 
 test('should send cookies from a request options when the key is duplicated in both a request options and CookieJar', async (t) => {
   const jar = new CookieJar();
-  const agent = new MixedCookieAgent({ jar });
+  const agent = new HttpCookieAgent({ jar });
 
   const { port } = await createTestServer([
     (req, res) => {
@@ -131,12 +105,35 @@ test('should send cookies from a request options when the key is duplicated in b
 
   await jar.setCookie('key=notexpected', `http://localhost:${port}`);
 
-  const { promise } = request(`http://localhost:${port}`, {
+  await urllib.request(`http://localhost:${port}`, {
     agent,
     headers: { Cookie: 'key=expected' },
-    method: 'GET',
   });
-  await promise;
+
+  t.plan(1);
+});
+
+test('should send cookies from the first response when redirecting', async (t) => {
+  const jar = new CookieJar();
+  const agent = new HttpCookieAgent({ jar });
+
+  const { port } = await createTestServer([
+    (_req, res) => {
+      res.statusCode = 301;
+      res.setHeader('Location', '/redirect');
+      res.setHeader('Set-Cookie', 'key=value');
+      res.end();
+    },
+    (req, res) => {
+      t.is(req.headers['cookie'], 'key=value');
+      res.end();
+    },
+  ]);
+
+  await urllib.request(`http://localhost:${port}`, {
+    agent,
+    followRedirect: true,
+  });
 
   t.plan(1);
 });
@@ -146,7 +143,7 @@ test('should emit error when CookieJar#getCookies throws error.', async (t) => {
   jar.getCookies = async () => {
     throw new Error('Error');
   };
-  const agent = new MixedCookieAgent({ jar });
+  const agent = new HttpCookieAgent({ jar });
 
   const { port } = await createTestServer([
     (_req, res) => {
@@ -155,11 +152,11 @@ test('should emit error when CookieJar#getCookies throws error.', async (t) => {
     },
   ]);
 
-  const { promise } = request(`http://localhost:${port}`, {
-    agent,
-    method: 'GET',
+  await t.throwsAsync(() => {
+    return urllib.request(`http://localhost:${port}`, {
+      agent,
+    });
   });
-  await t.throwsAsync(() => promise);
 
   t.plan(1);
 });
@@ -169,7 +166,7 @@ test('should emit error when CookieJar#setCookie throws error.', async (t) => {
   jar.setCookie = async () => {
     throw new Error('Error');
   };
-  const agent = new MixedCookieAgent({ jar });
+  const agent = new HttpCookieAgent({ jar });
 
   const { port } = await createTestServer([
     (_req, res) => {
@@ -178,11 +175,40 @@ test('should emit error when CookieJar#setCookie throws error.', async (t) => {
     },
   ]);
 
-  const { promise } = request(`http://localhost:${port}`, {
-    agent,
-    method: 'GET',
+  await t.throwsAsync(() => {
+    return urllib.request(`http://localhost:${port}`, {
+      agent,
+    });
   });
-  await t.throwsAsync(() => promise);
 
   t.plan(1);
+});
+
+test('should send post data when keepalive is enabled', async (t) => {
+  const times = 2;
+
+  const jar = new CookieJar();
+  const agent = new HttpCookieAgent({ jar, keepAlive: true });
+
+  const { port } = await createTestServer(
+    Array.from({ length: times }, (_, idx) => {
+      return async (req, res) => {
+        t.is(await readStream(req), `{ "index": "${idx}" }`);
+        t.is(req.headers['cookie'], 'key=expected');
+        res.end();
+      };
+    }),
+  );
+
+  await jar.setCookie('key=expected', `http://localhost:${port}`);
+
+  for (let idx = 0; idx < times; idx++) {
+    await urllib.request(`http://localhost:${port}`, {
+      agent,
+      data: `{ "index": "${idx}" }`,
+      method: 'POST',
+    });
+  }
+
+  t.plan(times * 2);
 });
