+++ src/__tests__/wreck.spec.mts
@@ -1,12 +1,10 @@
-import { promisify } from 'node:util';
-
+import Wreck from '@hapi/wreck';
 import test from 'ava';
-import request from 'request';
 import { CookieJar } from 'tough-cookie';
 
-import { HttpCookieAgent } from '../';
+import { HttpCookieAgent } from '../index.js';
 
-import { createTestServer, readStream } from './helpers';
+import { createTestServer, readStream } from './helpers.mjs';
 
 test('should set cookies to CookieJar from Set-Cookie header', async (t) => {
   const jar = new CookieJar();
@@ -19,9 +17,8 @@ test('should set cookies to CookieJar from Set-Cookie header', async (t) => {
     },
   ]);
 
-  await promisify(request)({
+  await Wreck.get(`http://localhost:${port}`, {
     agent,
-    url: `http://localhost:${port}`,
   });
 
   const cookies = await jar.getCookies(`http://localhost:${port}`);
@@ -42,9 +39,8 @@ test('should set cookies to CookieJar from multiple Set-Cookie headers', async (
     },
   ]);
 
-  await promisify(request)({
+  await Wreck.get(`http://localhost:${port}`, {
     agent,
-    url: `http://localhost:${port}`,
   });
 
   const cookies = await jar.getCookies(`http://localhost:${port}`);
@@ -68,9 +64,8 @@ test('should send cookies from CookieJar', async (t) => {
 
   await jar.setCookie('key=value', `http://localhost:${port}`);
 
-  await promisify(request)({
+  await Wreck.get(`http://localhost:${port}`, {
     agent,
-    url: `http://localhost:${port}`,
   });
 
   t.plan(1);
@@ -89,10 +84,9 @@ test('should send cookies from both a request options and CookieJar', async (t) 
 
   await jar.setCookie('key1=value1', `http://localhost:${port}`);
 
-  await promisify(request)({
+  await Wreck.get(`http://localhost:${port}`, {
     agent,
     headers: { Cookie: 'key2=value2' },
-    url: `http://localhost:${port}`,
   });
 
   t.plan(1);
@@ -111,10 +105,9 @@ test('should send cookies from a request options when the key is duplicated in b
 
   await jar.setCookie('key=notexpected', `http://localhost:${port}`);
 
-  await promisify(request)({
+  await Wreck.get(`http://localhost:${port}`, {
     agent,
     headers: { Cookie: 'key=expected' },
-    url: `http://localhost:${port}`,
   });
 
   t.plan(1);
@@ -137,9 +130,9 @@ test('should send cookies from the first response when redirecting', async (t) =
     },
   ]);
 
-  await promisify(request)({
+  await Wreck.get(`http://localhost:${port}`, {
     agent,
-    url: `http://localhost:${port}`,
+    redirects: 1,
   });
 
   t.plan(1);
@@ -160,9 +153,8 @@ test('should emit error when CookieJar#getCookies throws error.', async (t) => {
   ]);
 
   await t.throwsAsync(() => {
-    return promisify(request)({
+    return Wreck.get(`http://localhost:${port}`, {
       agent,
-      url: `http://localhost:${port}`,
     });
   });
 
@@ -184,9 +176,8 @@ test('should emit error when CookieJar#setCookie throws error.', async (t) => {
   ]);
 
   await t.throwsAsync(async () => {
-    return promisify(request)({
+    return Wreck.get(`http://localhost:${port}`, {
       agent,
-      url: `http://localhost:${port}`,
     });
   });
 
@@ -212,11 +203,9 @@ test('should send post data when keepalive is enabled', async (t) => {
   await jar.setCookie('key=expected', `http://localhost:${port}`);
 
   for (let idx = 0; idx < times; idx++) {
-    await promisify(request)({
+    await Wreck.post(`http://localhost:${port}`, {
       agent,
-      body: `{ "index": "${idx}" }`,
-      method: 'POST',
-      url: `http://localhost:${port}`,
+      payload: `{ "index": "${idx}" }`,
     });
   }
